scripts: fixdep: update to Linux v6.4-rc4
Some other files required changes too because cmd_$@ was renamed to
savedcmd_$@ in *.cmd files.

This works around an issues that appears when trying to compile barebox
with clang. The issue manifests on compiling device tree files and leads
to corrupted .cmd files, breaking recompilation until those are deleted.
This issue seem to appear because of the way we invoke the compiler with
-MT option argument for some reason not fully replacing the target name
in the generated dependency file, but instead simply appending it to the
list of targets. The changes in fixdep upstream improved the parsing
capabilities of fixdep, which now successfully handles multiple target
rules, skipping all but the last target.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>