MIPS: add qemu-malta64el_defconfig
This defconfig will compile barebox for Malta board and MIPS64 5KEc/5KEf
CPUs, emulatable in QEMU. The analogous 32-bit defconfig uses big endian
CPU as the default one. So, for the sake of the variety, and to improve
the coverage of options, set the CPU as little endian for this one.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>